+++ CHANGELOG.md
@@ -8,8 +8,8 @@
 
 - Sync `tags`, `extra`, `fingerprint`, `level`, `environment` and `breadcrumbs` from `sentry-cocoa` during event processing. ([#2713](https://github.com/getsentry/sentry-react-native/pull/2713))
   - `breadcrumb.level` value `log` is transformed to `debug` when syncing with native layers.
-  - Deprecated `breadcrumb.level` value `critical` is removed and not transformed.
-  - Default `breadcrumb.level` is `error`
+  - Remove `breadcrumb.level` value `critical` transformation to `fatal`.
+  - Default `breadcrumb.level` is `info`
 
 ### Breaking changes